New states for log_normal.ml.
Now work with <m> and sigma_m as the state space for lognormal MCMC.
The priors are uniform in this space, and the same as used for the
exponential distribution: sigma_m^max(<m>) = (M_max - <m>)/2.  Thus,
the maximum mass is always kept 2 sigma away from the mean mass.